Use typescript module augmentation to declare custom mixin
Fixes #41.
